+++ lib/src/MoveGeneration.cpp
@@ -6,7 +6,7 @@
 #include "../include/Game.h"
 using namespace std;
 
-vector<Move> Game::moves(int square) {
+vector<Move> Game::moves(uint8_t square) {
 	vector<Move> moves;
 	addMoves(square, moves);
 
@@ -21,7 +21,7 @@ vector<Move> Game::moves(int square) {
 
 vector<Move> Game::moves() {
 	vector<Move> moves;
-	for (int i = 0; i < 64; i++) {
+	for (uint8_t i = 0; i < 64; i++) {
 		if (pieces[i] & turn) {
 			addMoves(i, moves);
 		}
@@ -37,14 +37,14 @@ vector<Move> Game::moves() {
 }
 
 bool Game::isLegal(Move m) {
-	int mine = turn;
+	uint8_t mine = turn;
 	make(m);
 	bool check = checked(mine);
 	undo();
 	return !check;
 }
 
-bool Game::checked(int color) {
+bool Game::checked(uint8_t color) {
 	for (Move a : attacked(color)) {
 		if (pieces[a.to] == (KING | color)) {
 			return true;
@@ -53,8 +53,8 @@ bool Game::checked(int color) {
 	return false;
 }
 
-void Game::addMoves(int square, vector<Move>& moves) {
-	int type = pieces[square] & TYPE;
+void Game::addMoves(uint8_t square, vector<Move>& moves) {
+	uint8_t type = pieces[square] & TYPE;
 	if (type == PAWN) pawnMoves(square, moves);
 	if (type == KNIGHT) knightMoves(square, moves);
 	if (type == BISHOP) bishopMoves(square, moves);
@@ -63,11 +63,11 @@ void Game::addMoves(int square, vector<Move>& moves) {
 	if (type == KING) kingMoves(square, moves);
 }
 
-void Game::pawnMoves(int square, vector<Move>& moves) {
-	int piece = pieces[square];
+void Game::pawnMoves(uint8_t square, vector<Move>& moves) {
+	uint8_t piece = pieces[square];
 
 	int step = 1;
-	int startRow = 1;
+	uint8_t startRow = 1;
 	if (piece & WHITE) {
 		step = -1;
 		startRow = 6;
@@ -75,23 +75,23 @@ void Game::pawnMoves(int square, vector<Move>& moves) {
 
 	addPawnMove(square, moves, step, 0, false);
 
-	int row = square / 8;
-	int one = square + step * 8;
+	uint8_t row = square / 8;
+	uint8_t one = square + step * 8;
 	if (row == startRow && pieces[one] == EMPTY) {
 		addPawnMove(square, moves, step * 2, 0, false);
 	}
 
 	for (int d : {-1, 1}) {
-		int to = square + step * 8 + d;
+		uint8_t to = square + step * 8 + d;
 		if (pieces[to] != EMPTY || to == enPassant) {
 			addPawnMove(square, moves, step, d);
 		}
 	}
 }
 
-void Game::addPawnMove(int from, vector<Move>& moves, int r, int c,
+void Game::addPawnMove(uint8_t from, vector<Move>& moves, uint8_t r, uint8_t c,
 											 bool capture) {
-	int row = from / 8 + r;
+	uint8_t row = from / 8 + r;
 	if (row == 0 || row == 7) {
 		addJump(from, moves, r, c, capture, Game::QUEEN);
 		addJump(from, moves, r, c, capture, Game::ROOK);
@@ -102,7 +102,7 @@ void Game::addPawnMove(int from, vector<Move>& moves, int r, int c,
 	}
 }
 
-void Game::knightMoves(int square, vector<Move>& moves) {
+void Game::knightMoves(uint8_t square, vector<Move>& moves) {
 	addJump(square, moves, -2, -1);
 	addJump(square, moves, -2, 1);
 	addJump(square, moves, -1, -2);
@@ -113,26 +113,26 @@ void Game::knightMoves(int square, vector<Move>& moves) {
 	addJump(square, moves, 2, 1);
 }
 
-void Game::bishopMoves(int square, vector<Move>& moves) {
+void Game::bishopMoves(uint8_t square, vector<Move>& moves) {
 	addSlide(square, moves, -1, -1);
 	addSlide(square, moves, -1, 1);
 	addSlide(square, moves, 1, -1);
 	addSlide(square, moves, 1, 1);
 }
 
-void Game::rookMoves(int square, vector<Move>& moves) {
+void Game::rookMoves(uint8_t square, vector<Move>& moves) {
 	addSlide(square, moves, -1, 0);
 	addSlide(square, moves, 0, -1);
 	addSlide(square, moves, 0, 1);
 	addSlide(square, moves, 1, 0);
 }
 
-void Game::queenMoves(int square, vector<Move>& moves) {
+void Game::queenMoves(uint8_t square, vector<Move>& moves) {
 	rookMoves(square, moves);
 	bishopMoves(square, moves);
 }
 
-void Game::kingMoves(int square, vector<Move>& moves) {
+void Game::kingMoves(uint8_t square, vector<Move>& moves) {
 	addJump(square, moves, -1, -1);
 	addJump(square, moves, -1, 0);
 	addJump(square, moves, -1, 1);
@@ -161,9 +161,9 @@ void Game::kingMoves(int square, vector<Move>& moves) {
 	}
 }
 
-bool Game::canCastle(int color, int side) {
-	int castle = castleBlack;
-	int king = 4;
+bool Game::canCastle(uint8_t color, uint8_t side) {
+	uint8_t castle = castleBlack;
+	uint8_t king = 4;
 	if (color == WHITE) {
 		castle = castleWhite;
 		king = 60;
@@ -171,14 +171,21 @@ bool Game::canCastle(int color, int side) {
 
 	if (!(castle & side)) return false;
 
-	vector<int> gap = {};
+	vector<uint8_t> gap = {};
 	if (side == QUEEN) {
-		gap = {king - 1, king - 2, king - 3};
+		gap = {
+				static_cast<uint8_t>(king - 1),
+				static_cast<uint8_t>(king - 2),
+				static_cast<uint8_t>(king - 3),
+		};
 	} else {
-		gap = {king + 1, king + 2};
+		gap = {
+				static_cast<uint8_t>(king + 1),
+				static_cast<uint8_t>(king + 2),
+		};
 	}
 
-	for (int g : gap) {
+	for (uint8_t g : gap) {
 		if (pieces[g] != EMPTY) return false;
 	}
 
@@ -190,48 +197,48 @@ bool Game::canCastle(int color, int side) {
 	return true;
 }
 
-vector<Move> Game::attacked(int color) {
+vector<Move> Game::attacked(uint8_t color) {
 	vector<Move> attacks;
 
-	for (int i = 0; i < 64; i++) {
+	for (uint8_t i = 0; i < 64; i++) {
 		if (castleBlack && i == 4) continue;
 		if (castleWhite && i == 60) continue;
 
-		int attacker = pieces[i];
+		uint8_t attacker = pieces[i];
 		if (!attacker) continue;
 		if (attacker & color) continue;
 
 		addMoves(i, attacks);
 
 		if (attacker == (WHITE | PAWN) && i / 8 == 1) {
-			attacks.push_back(Move{i, i - 7});
-			attacks.push_back(Move{i, i - 9});
+			attacks.push_back(Move{i, static_cast<uint8_t>(i - 7)});
+			attacks.push_back(Move{i, static_cast<uint8_t>(i - 9)});
 		}
 		if (attacker == (BLACK | PAWN) && i / 8 == 6) {
-			attacks.push_back(Move{i, i + 7});
-			attacks.push_back(Move{i, i + 9});
+			attacks.push_back(Move{i, static_cast<uint8_t>(i + 7)});
+			attacks.push_back(Move{i, static_cast<uint8_t>(i + 9)});
 		}
 	}
 
 	return attacks;
 }
 
-void Game::addSlide(int from, vector<Move>& moves, int r, int c) {
-	for (int i = 1; i < 8; i++) {
+void Game::addSlide(uint8_t from, vector<Move>& moves, int r, int c) {
+	for (uint8_t i = 1; i < 8; i++) {
 		if (!addJump(from, moves, r * i, c * i)) {
 			return;
 		}
 	}
 }
 
-bool Game::addJump(int from, vector<Move>& moves, int r, int c, bool capture,
-									 int promote) {
-	int row = from / 8 + r;
+bool Game::addJump(uint8_t from, vector<Move>& moves, int r, int c,
+									 bool capture, uint8_t promote) {
+	uint8_t row = from / 8 + r;
 	if (row < 0 || row > 7) return false;
-	int col = from % 8 + c;
+	uint8_t col = from % 8 + c;
 	if (col < 0 || col > 7) return false;
 
-	int to = row * 8 + col;
+	uint8_t to = row * 8 + col;
 	if ((pieces[to] & COLOR) == (pieces[from] & COLOR)) return false;
 
 	if (pieces[to]) {
